fix: type conversion issue

+++ client/android/app/src/main/kotlin/org/calypsonet/keyple/demo/reload/remote/ui/ChargeActivity.kt
@@ -125,7 +125,7 @@ class ChargeActivity : AbstractCardActivity() {
         val ticketToBeLoaded = intent.getIntExtra(SelectTicketsActivity.TICKETS_NUMBER, 0)
         writeContractInputDto.contractTariff =
             PriorityCode.findEnumByKey(
-                intent.getByteExtra(SelectTicketsActivity.SELECTED_TICKET_PRIORITY_CODE, 0).toInt())
+                intent.getIntExtra(SelectTicketsActivity.SELECTED_TICKET_PRIORITY_CODE, 0))
         writeContractInputDto.ticketToLoad = ticketToBeLoaded
 
         val writeTitleOutput =
